Submitting questions It works, question and answer check is build in. BUT to achieve this I had to break a bit of code. Missing an awnser doesn't put the selected anwser back to null.
Some other code should also be rewitten to make it better. Hitting the question should be (and is now partially) handled in the question object and not in the bullet class.

The given answer should be saved within the question and is received from the bullet right now.  If answer = null, question cant get the anwser in time so thats why I broke a bit of the code.